Fixing float / double conversion

+++ suripu-core/src/main/java/com/hello/suripu/core/util/TimelineUtils.java
@@ -465,8 +465,8 @@ public class TimelineUtils {
      */
     public static String generateMessage(final SleepStats sleepStats) {
         final Integer percentageOfSoundSleep = Math.round(new Float(sleepStats.soundSleepDurationInMinutes) /sleepStats.sleepDurationInMinutes * 100);
-        final float sleepDurationInHours = sleepStats.sleepDurationInMinutes / 60.0;
-        final float soundDurationInHours = sleepStats.soundSleepDurationInMinutes / 60.0;
+        final double sleepDurationInHours = sleepStats.sleepDurationInMinutes / 60.0;
+        final double soundDurationInHours = sleepStats.soundSleepDurationInMinutes / 60.0;
         return String.format("You slept for a total of **%.1f hours**, soundly for %.1f hours, (%d%%) and moved %d times",
                 sleepDurationInHours, soundDurationInHours, percentageOfSoundSleep, sleepStats.numberOfMotionEvents);
     }
